test: split native backend suites
